--- a/Plugins/RTSFormations/Content/DA_AgentFormation.uasset
+++ b/Plugins/RTSFormations/Content/DA_AgentFormation.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:25b9bef364eb89ab5f1afc55d48473f632ed62f6dfd396214d24489c128fae28
+size 3897