--- a/Plugins/RTSFormations/Content/DA_AgentFormation.uasset
+++ b/Plugins/RTSFormations/Content/DA_AgentFormation.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:25b9bef364eb89ab5f1afc55d48473f632ed62f6dfd396214d24489c128fae28
-size 3897
+oid sha256:e09aeb6d360f2f876a8db14199a67c0c2c326abc3639ca6dacdb7d867c729095
+size 5886