--- a/Plugins/RTSFormations/Content/RTSFormationsExample.umap
+++ b/Plugins/RTSFormations/Content/RTSFormationsExample.umap
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:13e8cbdf14a10979f5ffbfae0622ac95f1b54bc0c81546434399e5a405ca4bfd
+size 36392