--- a/Plugins/RTSFormations/Content/RTSFormationsExample.umap
+++ b/Plugins/RTSFormations/Content/RTSFormationsExample.umap
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:13e8cbdf14a10979f5ffbfae0622ac95f1b54bc0c81546434399e5a405ca4bfd
-size 36392
+oid sha256:027894da6515631e5c3356ec6cab86bf95505056734fdedf77b3107af446718a
+size 39665