--- a/Plugins/RTSFormations/RTSFormations.uplugin
+++ b/Plugins/RTSFormations/RTSFormations.uplugin
@@ -1,0 +1,24 @@
+{
+	"FileVersion": 3,
+	"Version": 1,
+	"VersionName": "1.0",
+	"FriendlyName": "RTSFormations",
+	"Description": "",
+	"Category": "Other",
+	"CreatedBy": "JiRath",
+	"CreatedByURL": "",
+	"DocsURL": "",
+	"MarketplaceURL": "",
+	"SupportURL": "",
+	"CanContainContent": true,
+	"IsBetaVersion": true,
+	"IsExperimentalVersion": false,
+	"Installed": false,
+	"Modules": [
+		{
+			"Name": "RTSFormations",
+			"Type": "Runtime",
+			"LoadingPhase": "Default"
+		}
+	]
+}