--- a/Plugins/RTSFormations/Resources/Icon128.png
+++ b/Plugins/RTSFormations/Resources/Icon128.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:f7239efaeefbd82de33ebe18518e50de075ea4188a468a9e4991396433d2275f
+size 12699