--- a/Plugins/RTSFormations/Source/RTSFormations/Private/RTSFormationSubsystem.cpp
+++ b/Plugins/RTSFormations/Source/RTSFormations/Private/RTSFormationSubsystem.cpp
@@ -12,7 +12,8 @@ void URTSFormationSubsystem::DestroyEntity(UMassAgentComponent* Entity)
 {
 	UMassEntitySubsystem* EntitySubsystem = GetWorld()->GetSubsystem<UMassEntitySubsystem>();
 	EntitySubsystem->Defer().DestroyEntity(Entity->GetEntityHandle());
-	UE_LOG(LogTemp, Error, TEXT("Entity: %d"), Entity->GetEntityHandle().Index)
+	
+	// My current observer implementation doesnt handle entity destruction properly, so the logic is performed here for the time
 	Units.Remove(Entity->GetEntityHandle());
 	GetWorld()->GetSubsystem<UMassSignalSubsystem>()->SignalEntities(FormationUpdated, Units);
 }