--- a/Plugins/RTSFormations/Source/RTSFormations/Public/RTSFormations.h
+++ b/Plugins/RTSFormations/Source/RTSFormations/Public/RTSFormations.h
@@ -1,0 +1,15 @@
+// Copyright Epic Games, Inc. All Rights Reserved.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Modules/ModuleManager.h"
+
+class FRTSFormationsModule : public IModuleInterface
+{
+public:
+
+	/** IModuleInterface implementation */
+	virtual void StartupModule() override;
+	virtual void ShutdownModule() override;
+};