Early example of simple unit formations. There is still a lot of cleanup to be done but the core functionality is there.

- 'Rectangle' formation by specifying a `FormationLength` in the `RTSFormationTrait`
- Spawning and shooting entities will appropriately shift the formation to fill any gaps
- Encapsulated in a plugin (though it does depend on some things in the project)


![image](https://user-images.githubusercontent.com/1747157/178162688-83f6aef6-4331-46f7-b643-74eba72c35ce.png)

https://user-images.githubusercontent.com/1747157/178162858-c0dace7c-4474-4719-8cee-82d1a091d3d5.mp4


https://user-images.githubusercontent.com/1747157/178162889-841d7e2e-a90d-4d9f-816f-83684432b1e7.mp4


https://user-images.githubusercontent.com/1747157/178162918-2471bb9d-351c-4b40-bf5b-9be2541f01a0.mp4


Known issues
- When spawning a large amount of entities (>100) actor visualization appears to stop functioning. May be unrelated to my code.
- Performance was not a priority so some things can definitely be optimized
- `URTSFormationInitializer` does not properly handle entity destruction. It is temporarily handled in the subsystem.
- A few processors are bundled together. For readability, I might separate them in the future.
- Unit movement is not implemented but should theoretically be straightforward.
- When units are destroyed, the entity index is sometimes not properly handled.