@Megafunk 2022-07-11
_When spawning a large amount of entities (>100) actor visualization appears to stop functioning. May be unrelated to my code._

The visual quantity limit is probably the in the LOD parameter config. I'll definitely look at this soon, still shoveling the rest of the project into plugins.