@Ji-Rath 2022-07-11
I finally tackled performance and gave into using arrays for formations.

With 5000 entities
Before
- RTSFormationInitializer - 100.1 ms
- RTSFormationUpdate 10 s

After
- RTSFormationInitializer - 378 μs
- RTSFormationUpdate - 22.2 ms

Kind of embarrassing looking at the original stats but still progress! I have ideas to make the update faster (currently only runs when an entity needs a new position), but I think it is 'good enough' for now.

https://user-images.githubusercontent.com/1747157/178365202-8f175609-96d8-4a98-a3f2-1ab6dc133e80.mp4