@Ji-Rath 2022-07-14
This is gonna be just a little messy, I made the mistake of creating this plugin in main, so Im just going to create a new PR with the new branch so I can do some more work